pyproject.toml
[project]
name = "travelserver"
version = "0.1.0"
description = "A simple MCP server for flight searches using Duffel API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp",
    "duffel-api",
    "pydantic>=2.0.0",
    "python-dotenv",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-include = ["server.py"]
